move fetches to saga
