+++ src/features/browserSlice.js
@@ -11,6 +11,7 @@ const browserSlice = createSlice({
   },
   reducers: {
     fetchSearchResult: () => {},
+    fetchPopular: () => {},
     setData: (state, { payload: data }) => {
       state.data = data;
     },
@@ -29,7 +30,15 @@ const browserSlice = createSlice({
   },
 });
 
-export const { setData, setStatus, setGenres, setPopularPeople, setPerson, fetchSearchResult } = browserSlice.actions;
+export const {
+  setData,
+  setStatus,
+  setGenres,
+  setPopularPeople,
+  setPerson,
+  fetchSearchResult,
+  fetchPopular,
+} = browserSlice.actions;
 
 export const selectBrowserState = (state) => state.browser;
 export const selectStatus = (state) => selectBrowserState(state).status;
@@ -37,6 +46,5 @@ export const selectData = (state) => selectBrowserState(state).data;
 export const selectGenres = (state) => selectBrowserState(state).genres;
 export const selectPopularPeople = (state) => selectBrowserState(state).people;
 export const selectPerson = (state) => selectBrowserState(state).person;
-
 
 export default browserSlice.reducer;
